requirements.txt
streamlit
streamlit-gpt-vis
fastapi
uvicorn[standard]
langchain
langchain-openai
langgraph
pandas
openpyxl
python-dotenv
httpx
python-multipart
